+++ themes/WebGuitarPro_AngularJs/assets/js/Player/tablature-player/player-controls/play-pause-button-item/play-pause-button-item.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { PlayPauseButtonItemComponent } from './play-pause-button-item.component';
+
+describe('PlayPauseButtonItemComponent', () => {
+  let component: PlayPauseButtonItemComponent;
+  let fixture: ComponentFixture<PlayPauseButtonItemComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ PlayPauseButtonItemComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(PlayPauseButtonItemComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});