+++ themes/WebGuitarPro_AngularJs/assets/js/pages/filemanager-file-upload.js
@@ -1,0 +1,6 @@
+require( '../includes/file-input.js' );
+
+$( function()
+{
+    
+});