+++ themes/WebGuitarPro_AngularJs/webpack.config.js
@@ -1,4 +1,5 @@
 const Encore = require('@symfony/webpack-encore');
+const AngularCompilerPlugin = require('@ngtools/webpack').AngularWebpackPlugin;
 
 Encore
     .setOutputPath( 'public/shared_assets/build/web-guitar-pro-angularjs/' )
@@ -12,6 +13,17 @@ Encore
     .enableSassLoader(function(sassOptions) {}, {
         resolveUrlLoader: true
     })
+    
+    /**
+     * Configure Angular Compiler and Loader
+     */
+    .enableTypeScriptLoader()
+    .addPlugin(new AngularCompilerPlugin({
+        "tsConfigPath": './themes/WebGuitarPro_AngularJs/assets/js/Player/tsconfig.app.json',
+        "entryModule": './themes/WebGuitarPro_AngularJs/assets/js/Player/player.module',
+        "skipCodeGeneration": true,
+        "exclude": [],
+    }))
     
     /**
      * Add Entries
@@ -36,11 +48,33 @@ Encore
      })
 
     
-    .addEntry('app', './themes/WebGuitarPro_AngularJs/assets/js/app.js')
-    .addEntry('tablature-player', './themes/WebGuitarPro_AngularJs/assets/js/pages/tablature-player.js')
+    /*
+     * ENTRY CONFIG
+     *
+     * Each entry will result in one JavaScript file (e.g. app.js)
+     * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
+     */
+    .addEntry('app', './themes/WebGuitarPro_AngularJs/assets/app.js')
+    .addEntry('tablature-player', './themes/WebGuitarPro_AngularJs/assets/js/Player/main.ts')
+    
+    .addEntry('authentication', './themes/WebGuitarPro_AngularJs/assets/js/pages/authentication.js')
+    .addEntry('tablatures', './themes/WebGuitarPro_AngularJs/assets/js/pages/tablatures.js')
+    .addEntry('tablatures-ext', './themes/WebGuitarPro_AngularJs/assets/js/pages/tablatures-ext.js')
+    .addEntry('tablature-edit', './themes/WebGuitarPro_AngularJs/assets/js/pages/tablature-edit.js')
+    
+    .addEntry( 'js/filemanager-file-upload', './themes/WebGuitarPro_AngularJs/assets/js/pages/filemanager-file-upload.js' )
+    .addEntry( 'js/profile', './themes/WebGuitarPro_AngularJs/assets/js/pages/profile.js' )
+    
+    .addEntry( 'paymentForm', './themes/WebGuitarPro_AngularJs/assets/js/Stripe/paymentForm.js' )
+    
+    .addEntry( 'shared-tablatures', './themes/WebGuitarPro_AngularJs/assets/js/pages/shared-tablatures.js' )
+    .addEntry( 'my-shares', './themes/WebGuitarPro_AngularJs/assets/js/pages/my-shares.js' )
 ;
 
 const config = Encore.getWebpackConfig();
 config.name = 'WebGuitarPro_AngularJs';
+config.resolve = {
+    extensions: ['.ts']
+};
 
 module.exports = config;